refactor: move flags+config to new cli package
