+++ cli/flags.go
@@ -1,9 +1,8 @@
-package server
+package cli
 
 import (
 	"time"
 
-	"github.com/Layr-Labs/eigenda-proxy/store"
 	"github.com/urfave/cli/v2"
 
 	opservice "github.com/ethereum-optimism/optimism/op-service"
@@ -326,8 +325,6 @@ func init() {
 }
 
 type CLIConfig struct {
-	RedisCfg      store.RedisConfig
-	S3Config      store.S3Config
 	EigenDAConfig Config
 	MetricsCfg    opmetrics.CLIConfig
 }
@@ -335,10 +332,8 @@ type CLIConfig struct {
 func ReadCLIConfig(ctx *cli.Context) CLIConfig {
 	config := ReadConfig(ctx)
 	return CLIConfig{
-		RedisCfg:      config.RedisConfig,
 		EigenDAConfig: config,
 		MetricsCfg:    opmetrics.ReadCLIConfig(ctx),
-		S3Config:      config.S3Config,
 	}
 }
