+++ cmd/server/entrypoint.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	proxycli "github.com/Layr-Labs/eigenda-proxy/cli"
 	"github.com/Layr-Labs/eigenda-proxy/metrics"
 	"github.com/Layr-Labs/eigenda-proxy/server"
 	"github.com/urfave/cli/v2"
@@ -17,14 +18,12 @@ func StartProxySvr(cliCtx *cli.Context) error {
 	oplog.SetGlobalLogHandler(log.Handler())
 	log.Info("Starting EigenDA Proxy Server", "version", Version, "date", Date, "commit", Commit)
 
-	cfg := server.ReadCLIConfig(cliCtx)
+	cfg := proxycli.ReadCLIConfig(cliCtx)
 	if err := cfg.Check(); err != nil {
 		return err
 	}
 	ctx, ctxCancel := context.WithCancel(cliCtx.Context)
 	defer ctxCancel()
-
-	m := metrics.NewMetrics("default")
 
 	log.Info("Initializing EigenDA proxy server...")
 
@@ -32,7 +31,8 @@ func StartProxySvr(cliCtx *cli.Context) error {
 	if err != nil {
 		return fmt.Errorf("failed to create store: %w", err)
 	}
-	server := server.NewServer(cliCtx.String(server.ListenAddrFlagName), cliCtx.Int(server.PortFlagName), daRouter, log, m)
+	m := metrics.NewMetrics("default")
+	server := server.NewServer(cliCtx.String(proxycli.ListenAddrFlagName), cliCtx.Int(proxycli.PortFlagName), daRouter, log, m)
 
 	if err := server.Start(); err != nil {
 		return fmt.Errorf("failed to start the DA server: %w", err)
